Package cache improvements
New async operations:
- `add_url()`
- `copy_or_add()`
- `update_or_add()`

`add_url()` just adds the url, async.

`copy_or_add()` gets a package from the
cache, or from the url, if not in the cache,
and then it also adds it to the cache.

`update_or_add()` is similar, but always
tries to update the file from the url, using
an Etag. If an update is needed, it also updates
the cache.