Add a 'directpkg' column to the resulution result
This is TRUE if the package (not the ref) was directly
specified. It is needed, because installed:: is not
considered to be directly specified, that would break
the solver. `directpkg` makes sure that we know the
direct packages in the solution as well, which might
only include installed:: refs for some direct packages.