Fix $update() for any::
We need to get the version we planned for, otherwise
the dependencies or their versions might not be right.

This happened on GHA, where the cache had sessioninfo 1.2.x,
depending on cli 3.1.0. The plan had any::sessioninfo
and without RSPM it was resolved to sessioninfo 1.1.x,
and we also got cli 2.5.x.

The sesssioninfo 1.2.x in the cache was kept, and then it
failed with cli 2.5.x. Now we fix this by throwing away
sessioninfo 1.2.x in this case and installing 1.1.x.

Of course this is still bad because GHA now relies on
sessioninfo 1.2.x, and the sessioninfo call there does
not work. So we'll need to fix that as well.