Fix resolution if a package is not found
When a package is not found in the standard repos, we
added a `dep_types` column to the failed resolution, and
when merging the failed resolution with the successful
ones, they also got a `dep_types` column, which was `NA`.

Since there was a `dep_types` column already, the generic
resolution did not add one, so we ignored a lot of
dependencies in this case.

Closes #282.