Stop resolution multiplying rows from metadata (Fixes #379)
The metadata column was causing certain rows to recycle  over the
result data frame. The solution here is to check if the list is
not already length one and to wrap it into a list if it is not.

Also added an assertion to force the output into being one row
as desired.