New solver heuristic
To speed up solves with multiple repos containing
the same or slightly different versions of the same
packages, we now rule out any older source package that
has the same set of dependencies as the latest source
package.

We can rule these out safely, because they would surely not
be selected, anyway. It does not make sense to select an
older source package if it has the same exact dependencies
as a newer one.