Use source platform for non-native sysreqs lookup
If looking up sysreqs for another platform, and the platforms are
not explicitly specified, then set 'platforms' to source. Ideally we
would set it according to `config$sysreqs_platform`, but we don't
currently have a way of doing that. This simplification should not
fail currently, and even if we'll have sysreqs on non-source platforms
in the future, it should fail very rarely. Hopefully we can improve it.

Helps with #781.